Remove Composite Steps Runner => consolidate into Composite Steps Runner
